refactor: share radix and decorator reader
