+++ src/plugins/DecoratorPlugin.js
@@ -1,14 +1,4 @@
-export function TSDecoratorReader(stream, factory) {
-  const start = stream.getPosition();
-  if (stream.current() !== '@') return null;
-  let val = '@';
-  stream.advance();
-  while (stream.current() && /[A-Za-z0-9_$]/.test(stream.current())) {
-    val += stream.current();
-    stream.advance();
-  }
-  return factory('DECORATOR', val, start, stream.getPosition());
-}
+import { TSDecoratorReader } from './common/TSDecoratorReader.js';
 
 export const DecoratorPlugin = {
   modes: { default: [TSDecoratorReader] },
